0.2.0-alpha.9: MSVC coddling - phi to constant
MSVC really hates me.

+++ test/main.c
@@ -56,7 +56,7 @@ int main(int argc, char *argv[])
     vec3 teapotPos = {0.0, -1.5, 0.0};
     gm_setObjectPosition(teapot, teapotPos);
 
-	vec3 rotateRad = {0, M_PI / gm_getFrameRate(), 0};
+	vec3 rotateRad = {0, (float)(M_PI / (double)gm_getFrameRate()), 0};
 	while(true)
 	{
 		gm_beginFrame();
